OVS-DB: new OVS-DB methods in VTEP data client
This patch adds support for new OVS-DB method in the VTEP data client,
such as deletion of first and all MAC entries (multicast and unicast).

It also refactors the VTEP data client methods and tests to support
the new client methods and also replace some asserts that were used
for argument validation, yielding different test results depending
on whether asserting was enabled or not.

It includes the VtepMAC class for MAC addresses to be used by the VXGW
service, supporting both standard and non-standard addresses (e.g.
unknown-dst)

It replaces String arguments in the VtepDataClient with strong typed
arguments, removing unnecessary argument validation.

Change-Id: I9acb110a848270d659b3edac0a7ca3c309f62d6a
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
Signed-off-by: Galo Navarro <galo@midokura.com>