Make git submodules use https, remove midonet-docs
Using git:// for submodules requires setting up ssh keys both in the
client and github. This doesn't make much sense for our submodules which
are most of them public. This patch replaces the git proto with https in
all submodules (midonet-docs, odp-ovsdb - py client already used it).

Midonet-docs is removed, because it's not really maintained with the
code.

Change-Id: Ic29e0b665ee696efb1e4b48f273e2063a76bf36b
Signed-off-by: Galo Navarro <galo@midokura.com>