VXGW: Support for offline VTEP tunnel IP
Adds a tunnel IP entry to the ZooKeeper VTEP configuration data.
This change allows VTEP operations, such as creating a new
binding, even when a connection to the VTEP cannot be established.

The tunnel IP is added to the VTEP configuration when connecting
to the VTEP for the first time, either when the VTEP is configured
as part of the VXGW service, or when performing an API request that
results in a VTEP connection.

Ref: MN-2594

Change-Id: I58f3c3cbb9a5a72cc39a221807a4dff8eac898ee
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>