Bind services to the API module as singletons
This ensures that the MidoNet API module uses a single instance
of various services (e.g. REST API service, VXGW service, etc.),
such that the same instance is returned by Guice either when
starting or stopping a service. In this way, the stop method
is called on the appropriate instance in oreder to perform the
necesseary clean-up operations.

Ref: MN-2646

Change-Id: Ie762d8e1d7ad33e346b1361a95b5de4ed526efdd
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>