Bug-fix: VXGW service cannot set the L2 flooding proxy
The patch corrects the VXGW service, to create the correct
MacLocation object when setting the flooding proxy for a tunnel-zone.

The patch also fixes the computation of a tunnel-zone flooding
proxy host, removing null pointer exception warnings, and
updating a log message that would always display Integer.MIN_VALUE
as the weight of the selected flooding proxy.

Fixes: MN-2725
Change-Id: Idf8bb2353689de4590100a3ce9bfeb0a845f3160
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>