Use real Zookeeper in Midonet API tests
This patch replaces the MockDirectory infrastructure and uses a
TestingServer from the curator-test library to run against a real
Zookeeper. This required a significant amount of changes to remove the
previous mocking tricks. With this patch, the mocking instrumentation is
done completely outside of the code using standard Guice mechanisms.

Change-Id: I0f3698b4a23d0494357ccbdbef022c0265ce86d2
Signed-off-by: Galo Navarro <galo@midokura.com>