Modify port creations emit 404 when the device is invalid
This patch modifies the port creation on non-existing bridges emits 404.
The router resource handler did this but the bridge handler didn't. So
to fix this asynmetric behaviour, I added this logic to the bridge
handler.

This patch also adds the tests for the port creation on the routers and
bridges which don't exist. Previously this causes 500 because we assumed
the device ID is mandaroty. However now we don't force ports to have
their device IDs.

This fixes MN-1585.

Change-Id: I5445095f5745c8af8d1c2787b0196c287c44919d
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>