Move bridge VxlanPortId validation to API
The BridgeZkManager was performing some validations to ensure that users
don't modify the vxlan port through the API (because they should not).
This was imposing some extra scaffolding that is annoying for the
refactor leading to deprecate the dataclient and zkmanagers.

This patch adds a new validator for the Bridge dto that will ensure that
users never change the vxlan port id on a bridge. The underlying zk
managers don't anymore, and callers are responsible to ensure
correctness.

Change-Id: I3b6e269143f2e24d7c5bb5389c22c0fcc90debe4
Signed-off-by: Galo Navarro <galo@midokura.com>