Warns if a class is not registered for diagnosis
Currently if some class to be persisted is not registered with ZOOM,
ZOOM fails with a mysterious stack trace saying ClassNotFound: null.
With this patch, if a class is not registered on assertion, ZOOM logs
a WARN with the right class name.

Alternatively, isRegistered can throw a ClassNotFound Exception with
the right class name.

Change-Id: I6d3fbd9cbe9a26f1c86047fac2b3a0036c9fa44b