Test ZOOM.multi() in buik port/chain creation
Now the construction of a performance test topology can use multi
operation, based on the multi-operation size specified in the test
setup. For each bridge attached to a tenant router, the bridge itself
is normally created, but ports and rules can be created in bulk, up to
the max number of PersistentOps specifeid in the test setup.

Added a new test: "Multi-operation" can "bulk-create ports & chains".
- It tests two different multi-sizes, 100 and 1000.
- Creates a single tenant router and a single bridge, which has 1K
ports and 2K in/out rules.
- This test can pass, but the following test, "Max-ports-per-bridge"
can "be handled efficiently", building a 10K-ports bridge fails for
connection loss. Needs investigation.

This patch does NOT include:
- refactoring of ZoomStorageServiceTester
- refactoring of Devices to a Set.

Fix MN-2797
Change-Id: I2bd71aa2d26542dd52e6583dacb066bd67e331f2