ZOOM: Convert between Protobufs and POJOs using annotations
This patch includes a new annotation that support an automatic
conversion between POJOs and the Protobufs objects used by ZOOM.

To enhance a POJO with the automatic conversion capabilities,
it must extend the ZoomObject class, which exposes a constructor
with a Protobufs object argument for Protobufs to POJO conversion,
and a toProto() method for POJO to Protobufs conversion.

Fields to convert must receive a ZoomField annotation, which
specified the name of the corresponding Protobufs field, whether
the field is mandatory when reading from Protobufs, and whether
nulls are ignored when writing to Protobufs.

Change-Id: I2152a1dde0474f238a04dbedcce61d9a519cc9c7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>