Implement Zoom update validation.
Add an overload to StorageService.update() that takes an
UpdateValidator as a second argument. UpdateValidator has a single
function, validate(), that takes the new and old versions of the
object to be updated. StorageService.update() uses this as a callback
upon loading the old version of the object to allow callers to check
for invalid updates and/or copy state from the old to the new version
of the object.

Signed-off-by: Brandon Berg <bberg@midokura.com>
Change-Id: I38241196576b827bd71fb74b7403b24bad2462f3