Proto-to-Pojo conversion ignores unset Proto fields.
ZoomConvert currently throws an exception on an unset Protobuf field
that's associated to a POJO field in converting a Protobuf Message to
a POJO. However, this essentially means that a Protobuf field cannot
be unset (optional) if it is associated to a ZoomObject field, which
does not make sense. ZoomConvert, insead, should just ignore POJO
fields if the corresponding Protobuf fields are unset.

Change-Id: I2622b8d009ea7e61143e07e024566334c834a6fd