Re-architect the agent VTA to use ZOOM storage
This patch introduces a VirtualTopology (VT) classthat replaces
the VTA to store the virtual topology devices and to use the ZOOM
storage interface based on observables.

Each device type is managed by a DeviceManager, which stores a map
between the device ID and a DeviceObservable. The DeviceObservable
maps the high-level Protobuf device(s) to the low-level simulation
device, such that every update to a high-level device generates an
update to the simulation device. In addition, the DeviceObservable
stores a cached copy of the most recent version of the simulation
device, and makes it available via both a future and an observable.
It also provides a counter for the current subscriber to enable a
clean-up of unused simulation devices, if necessary.

Both the DeviceManager and DeviceObserver are abstract classes.
Implementations must specify how a simulation device is created
from the high-level Protobuf devices, and can attach specific
actions to perform for each update of a simulation device, such
as flow invalidation.

This patch demonstrates the implementation of a DeviceManager by
including PortManagerEx/PortObservable classes to manage ports.

Change-Id: Icda6346d8d590c5615fa4453ef1281fe23266294
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>