Re-architect the agent VTA to use ZOOM storage
This patch introduces a VirtualTopology (VT) classthat replaces
the VTA to store the virtual topology devices and to use the ZOOM
storage interface based on observables.

The VT receives updates for a device using a DeviceObservable
instance, which receives topology updates from the storage layer
and creates the corresponding simulation device. The device
observable exposes the device information as a future and reactive
observable. In addition, each device observable tracks the current
subscription to enable future garbage collection of unused observables
and updates a VT device cache that improves lookup performance for
existing devices.

The DeviceObseravable is an abstract class, and specific
implementations for each device type must specify how to combine
topology data read from storage, and create the corresponding
simulation device. The sub-classes can also override update methods
(onNext, onCompleted, onError), to take a specific action for a
given update, such as flow invalidation.

This patch demonstrates the implementation of a DeviceObservable for
a port device type: PortObservable.

Change-Id: Icda6346d8d590c5615fa4453ef1281fe23266294
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>