Make ZOOM asynchronous and remove HttpAwareZoom
The Storage service interface is made asynchronous to support the main
use cases for the RPC API. To this end, this patch:
- improves the get() and getAll() methods to use the instance cache, when
returning values;
- changes the subscribe() and subscribeAll() methods, to initialize
the instance/class caches asynchronously, without blocking the
subscribe methods.

HttpAwareZoom disappears since we're no longer using it due to the REST
API extermination. Also renaming StorageService to Storage for
conciseness.

Change-Id: Ied8517d4d8ed1f3ba9ee54db9f0bd3a8c30fb4da
Signed-off-by: Galo Navarro <galo@midokura.com>
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>