DatapathController.: move port resolution to host proxy
Move port resolution to HostRequestProxy, so that all event
handling in the datapath controller is not liable to non
existent ports or request timeouts, making it easier to
linearize operaitons and handle requets retries.

The HostRequestProxy will resolve ports into a new ResolvedHost
object. Ports that error out are filtered out. But, if any of them
does, a fresh request (unsubscribe/resubscribe) is sent to the VTPM,
which translates into a natural retry loop that doesn't mess with
the DatapathController event handling.

Change-Id: I9d1c0eedf48949ac9f689ecb4dd6ea950f728b81
Fixes: MN-2950
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>