Update patch set 1
Patch Set 1: Code-Review+1

Long-term we should either just point to the docs site in general (because the link supplied in this patch points to "v1.8" and will change in the future), or, we enhance the docs site to provide a "latest" URL that always points to the latest version of the documentation.

*For now* I think this patch is ok, in order to get the link in the README working asap.

I'll keep track of this issue, and update the link accordingly once the new AsciiDoc-based REST API document is in place and/or we have a generic URL available on the docs site.

Patch-set: 1
Reviewer: Jan Hilberath <1002753@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1