Removes the concept of PortSets
Removing port sets means that flooding a bridge is now a list
of individual to-port actions, which are processed at ingress
individually. The per-bridge tunnel key disappears. And so does
the processing of port filters at the egress hosts when a bridge
is flooded. The downside is that we now send a copy of the packet
per output port, instead of a copy per destination node.

This has a few consequences:

In bridges, PortSet tracking was the mechanism that made agent nodes
know which ports are active, because inactive port would only be candidates
for receiving traffic in the flooding case. Because port sets are now
gone, tracking of port status is now done in the ZK port object and
propagated all the way down into the simulation as an attribute of
the port object. Simulations filter inactive ports out.

Output filters are applied on the simulation layer. Flow translation
still needs to reorder output actions so that local ports are
processed before remote ones, so that encapsulation is applied
after forwarding to all local ports.

Change-Id: Ifaed132536b440dcced73d405681a8ca675bbe35
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>