Update patch set 4
Patch Set 4:

We had a planned blackout this morning and the configuration of the router seems to be gone due to that. So I can't connect to my machine at the office and I'm submitting the current WIP patch here to see the result of the gates. I'll finish it tomorrow at the office.

Patch-set: 4
Label: Code-Review=0