C3POStorageManager persists a transaction with Storage.
Provides a Storage implementaion of the C3PODatamanager service. It
receives a transaction of operations on external (most-likely Neutron)
models, translates them into equivalent MidoNet model operations, and
persist them with the Storage service.

Also defines a StorageException, which is a generic Exception class
to be a common super class for various Storage / Zoom related
exceptions.

NOTE: flushTopology() is not yet implemented.
MN-2999

Change-Id: I59f0a7a71c715102b25ff1ad567c26d25748e631