Clear references to non-existent chains
When add_router_interface is called with an existing
neutron port, which is the case when you specify IP
address for add interface operation in horizon,
we remove the port chains and convert it into logical
interface to connect to router. However we failed to
clear references to the port chains in the port data,
causing referential inconsistency.

Ref: MNA-1

Change-Id: Ie26fdf5f502561613e7f699c56b68286d9e6bdc9