Use Quagga FPM server feature
Quagga has introduced a new feature called the Fib Push Manager that
we can leverage in midonet. Currently, midonet listens to BGP for the
routing updates it sends to zebra by pretending to be zebra itself. We
can now instead just listen to updates from Zebra for changes in the FIB.

NOTES:
* We use netty instead of a select loop for handling the TCP connection.
* All of process/namespace management will be migrated to docker in the
  near future.

The changes to support this are:

MidolmanConfig.java: Added new config options for managing the zebra process
MidolmanActorsModule.java: got rid of the SelectServiceLoop that used to be
    needed for creating the zebra server channel
MidolmanModule.java: got rid of code creating the SelectServiceLoop.
SelectServiceLoop.java: deleted because no longer used.
RoutingHandler.java: Spawn the Zebra process alongside the BGPD process.
RoutingManagerActor.scala: No longer require the zebraLoop.
FpmServer.scala: The FPM packet deserializer and route installer.
ZebraConnection.scala: We no longer pretend to be zebra, so this file is no
    longer needed.
ZebraProcess.scala: A similar process to BGPd. Just a wrapper around the
    command line issued from java.
ZebraProtocol.scala: We no longer need this packet deserializer because
    we are not pretending to be zebra anymore.
ZebraProtocolHandler.scala: We no longer need this handler because we are
    not pretending to be zebra anymore.
ZebraServer.scala: We no longer need to pretend to be zebra.
MidolmanBenchmark.scala: no need to mock the zebra select loop anymore.

Change-Id: Ia95df58328df5b956d7842454bff525e543548dc
Signed-off-by: Joe Mills <joe@midokura.com>