ZOOM implements flush as version bumping.
This patch implements flush for ZOOM as data path version bump. The
data path node for class, for example, Bridge will be in the form of,

[BASE_PATH_PREFIX]/[VERSION_#]/Bridge

where base_path_prefix is the path prefix string specified on ZOOM
construction, and VERSION_# is monotonically incremented data path
version number, initially starting from 1.

A new ZOOM instance checks for the version number upon getting built.
If it finds one, it sets its data set version number to the found
value, so that in case of a cluster fail-over for example, a new
instance would be able to take over where the previous ZOOM instance
left off. In other words, in a cluster fail-over, the new leader needs
to initialize and build a new ZOOM instance upon taking over the
leadership.

Upon version number initialization, it sets a watcher to the version
number node, so that it'd be notified if another ZOOM instances bumps
the version number.

Fix MN-2999

Change-Id: Ibd5892a74ce5edb2c8b5f7542d78b6e7af328e6d