Update patch set 3
Patch Set 3:

(1 comment)

Thank you for the comment, Alex. This patch / test is primarily meant to be a form of documentation of the current ZOOM. I thought it was a good idea to record the current behaviour in the form of a unit test, but it's not critical whether this patch gets merged or not.

Patch-set: 3
Reviewer: Tomohiko Kimura <1002760@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0