Fix deadlock in ZookeeperObjectMapperTests
We believe the deadlocks are being caused by startGc()'s notify() call
happening before gcRunnable.run()'s wait() call. This change adds a
sentinel that should prevent a deadlock in that situation. If that's
the true cause, this should fix the deadlocks.

Update: That may or may not have been the cause of some deadlocks, but
some additional deadlocks were being caused by the Runnable closure
created in MockZookeeperObjectMapper.scheduleCacheGc() closing around
the old value of gcRunnable (from the prior test), causing the
synchronization not to work correctly. I broke it out into a regular
class so that gcRunnable will be bound by name rather than value.

Change-Id: I7b832dac4095e18c51cae2ed0e4186bd72ff95a8
Signed-off-by: Brandon Berg <bberg@midokura.com>