Metering: prepare FlowManager / FlowController
Updating meters upon flow deletion means flow metadata needs
to be around when the kernel confirms a flow deletion. This was
not the case before this patch, the flow manager lost track of
wildcard flows as soon as the requests to remove their dp flows
were queued.

This patch changes that. The removeFlow() call in the FlowManagerHelper
gets a pointer to the owning wildcard flow (nullable, there may be
none). The helper implementation in the flow controller keeps the
wildcard flow alive by incrementing the reference count and releasing
only after the flow has been deleted.

Change-Id: I245c1b3899b171a659ad1d7959f9a35e5f8ab2cf
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>