Modify FlowStateReplicator to keep packet size small
This patch modifies FlowStateReplicator to use the FlowState
message length instead of the fixed size, which was MTU -
GRE_ENCAPSULATION_OVERHEAD.

This modification assumes a FlowState message doesn't exceed
the MTU size. In addition to that, this patch copies the byte
array which range is specified dynamically everytime when
pushState is invoked.

This fixes another problem described in MN-3026.

Change-Id: I1fae9636fa1b501d670cb851ecc311e5a1ee4e6a
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>