Disables the subscription tests in ZOOMTests.
We found that testSubscribeWithGc and testSubscribeAllWithGc have some
race conditions, which may result in a thread waiting forever to be
waken up thus the Jenkins job getting stuck and timing out. We figured
out one race situation, but have been unable to provide a complete
solution. This patch disables the culprit tests till we have the
complete solution.

Change-Id: Id81de0f46fd1c248497a876ac229e0bcf17a2f09