Fix GC problems in ZookeeperObjectMapperTests
We believe the deadlocks are being caused by startGc()'s notify() call
happening before gcRunnable.run()'s wait() call. This change adds a sentinel
that should prevent a deadlock in that situation (thanks for fixing this
Brandon). This patch also ensures that the GC thread gets unblocked after
every test by calling startGc and waitForGc in the suite teardown.

This patch also fixes testSubscribe() by updating the bridge in zoom to add a
port to it.

Change-Id: Ida3bac6af6bfeb29a62f3392bc44b2201dad513b
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>