An instance observable for topology storage
Introduces an RX InstanceObservable, which wraps a subscription
to an instance observable, and where all updates emitted by the
storage layer are encapsulated by an Event object, which are
emitted both for observable completion and error. This allows
the aggregation of multiple instance observables, yet it supports
the notification of events such as instance deletion.

The observable also exposes a complete() method, which allows
an thrid-party caller to inject a completed notification,
unsubscribing all current subscribers.

In a future update, this observable may also provide storage error
handling.

Ref: MN-2986

Change-Id: I994852dec0579c02cbdd10ccb0578970d7a7cc45
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>