A virtual topology redirector
This patch introduces a virtual topology redirector for the
Midolman virtual topology actor, designed to intercept
request for virtual devices, and redirect them either to
the VirtualTopologyActor or to the VirtualTopology using the
new storage model.

To maximize compatibility with the current agent design, and
to minimize the required changes, the redirector is implemented
as a super class of the VirtualTopologyActor, is an AKKA actor,
and can intercept request in the actor's receive() methods.

Ref: MNA-103

Change-Id: Id84479609f0218a97b4ea76030654f93eeb7d2f7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>