A VTPM redirector
This patch introduces a VTPM redirector that allows to use the
VirtualToPhysicalMapper with the old and the new cluster design. A
boolean flag in ZookeeperConfig called cluster_storage_enabled tells
whether the new cluster should be used.

When the new cluster is used, we rely on the tunnel zone and host
mappers to obtain the corresponding simulation objects. When the old
cluster should be used, we use the host and tunnel zone  managers.

The VTPM redirector is a temporary class that will be removed once the
old cluster is permanently deleted.

Ref: MN-3125

Change-Id: I2e0df7a7490179bc6df04ba92c204484bc7a401d
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>