[WIP] A VTPM redirector
This patch introduces a VTPM redirector that allows to use the
VirtualToPhysicalMapper with the old and the new cluster design. A
boolean flag in ZookeeperConfig called cluster_storage_enabled tells
whether the new cluster design should be used.

For the moment, only tunnel zones can be obtained with the
VirtualToPhysicalMapper when the new cluster is enabled. A subsequent
patch will allow to obtain hosts as well.

The VTPM redirector is a temporary class that will be removed once the
old cluster is permanently deleted.

Change-Id: I2e0df7a7490179bc6df04ba92c204484bc7a401d
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>