ZOOM: Add ephemeral ownership support
Introduces a StorageWithOwnership trait with methods that
allow ownership-based storage operations, and the
implementation thereof for ZOOM storage.

The supported ownership modes are the following:
- Exclusive: an object can only have one owner, specified
at creation time, and which controls the update and delete
operations
- Shared: an object can zero or more owners, specified either
during the creation or the update of an object. An object can
only be deleted when all owners have released their ownership
of the object. Shared ownership is backwards compatible with
ZOOM operations that do not user owners.

In ZOOM, ownership for a stored object is specified by
one or more emphemeral children.

Change-Id: I4ff8df8ab5e5dc24bb2288a9407e058aef89b8d7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>