[WIP Adds ephemeral ownership support to ZOOM storage
Introduces a StorageWithOwnership trait with methods that
allow ownership-based storage operations, and the
implementation thereof for ZOOM storage.

The supported ownership modes are the following:
- None: does not use ownership and maintains the existing
functionality
- Exclusive: an object can only have one owner, specified
at creation time, and which controls the update and delete
operations
- Shared: an object can have multiple owners, specified either
during the creation or the update of an object. An object can
only be deleted when all owners have released their ownership
of the object.

In ZOOM, ownership for a stored object is specified by
one or more emphemeral children.

Change-Id: I4ff8df8ab5e5dc24bb2288a9407e058aef89b8d7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>