ObservablePathNodeCache closes child observables
When the cache is closed (typically when the ClassSubscriptionCache
decides to), the main observable was closed but it did not close the
children observables that it had emitted. This patch corrects this, so
that any subscribers that could have raced with the close() are always
aware that they need to request a new Observable.

The semantics of close() are consolidated now in both
ObservablePathChildrenCache and ObservableNodeCache, triggering an
onError on all Observables, signalling specifically a disconnection from
the backend, instead of the ambiguous onComplete (that insteads signals
a deletion).

This patch also adds some minor fixes pending from a previous review,
plus some bug fixes in connection management (and tests).

Change-Id: Ice43813a678cb12f2de34de98ea93fe3988d7f3e
Signed-off-by: Galo Navarro <galo@midokura.com>