Add MN Cluster implementation doc to specs
This document tries to offer context into the ongoing work to implement
a new topology propagation pipeline, aimed at a developer audience. It
includes basic context on the architecture changes proposed for 2015.02,
and tries to provide pointers into specific areas of the codebase that
are being evolved as part of this effort.

Change-Id: I1a3c3477d287c465bce8ba7e9a1704413fb0a1c8
Signed-off-by: Galo Navarro <galo@midokura.com>