Add MN Cluster implementation doc to specs
This document tries to offer context into the ongoing work to implement
the MidoNet 2.0 cluster, aimed at a developer audience. It includes
basic context on the architecture changes proposed for 2.0, and tries to
provide pointers into specific areas of the codebase that are being
evolved as part of this effort.

Change-Id: I1a3c3477d287c465bce8ba7e9a1704413fb0a1c8
Signed-off-by: Galo Navarro <galo@midokura.com>