Update patch set 1
Patch Set 1:

(1 comment)

Thanks for the writing, Galo. Just one comment: I think it is not clear form the text if the 'Translator' if composed of the 'Synchronizer' and the 'Importer' (that's my impression), or if it is meant to be a separate entity... Maybe it is worth to state that explicitly.

Patch-set: 1
Reviewer: Ernest Artiaga <1002750@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0