Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Mostly lgtm, the only thing is that I was referring to the TODOs in ZOOM regarding instance and class cache closing. It seems to me that subscribers will still not see their subscriptions complete after a flush. If you want to do that in another patch that's fine with me as well.

Patch-set: 2
Label: Code-Review=-1