Make VxlanGatewayService a ClusterMinion
This involves a refactor to remove the identity between cluster and vxgw
service, involving extracting the generation of the cluster node uuid
from the vxgw service so it's now done by a ClusterNode setup, or when
running embedded inside the API by the Jersey setup.

The patch keeps backwards compatibility with the midonet-api
configuration so no changes are required in existing deployments of the
API.

Note that the console runner still won't work since we need to move a
bunch of guice modules from midolman to cluster, which will be done in a
separate (and big) patch. The embedded cluster in the REST API will work
ok.

Change-Id: Ibed85561141a30e649e9ce6252abea7d5b4d6eab
Signed-off-by: Galo Navarro <galo@midokura.com>