Tests NeutronImporter with SQLite/ZK test server.
This patch runs an end-to-end test for NeutronImporter daemon with
SQLite as the sql DB and the in-memory ZooKeeper test server. It
starts the Cluster daemon from inside a Scala Test by providing a test
Minion configuration for NeutronImporter and running NeutronImporter
properly as a Minion. Sets up the MidoNet Tasks DB table on SQLite
with a mock DataSource adaptor for SQLite as SQLite doesn't seem to
provide JDBC 2.0 API. Uses Curator's in-memory ZooKeeper testing
server as a ZK backend.

Fix MN-2999
Change-Id: I57c966dee31214670793ddc601e7a8d5dba6f4e3