Update patch set 13
Patch Set 13:

(3 comments)

I think we can abandon this in favor of merge. The only difference is that merge waits for all observables to complete, before completing, but I think it's easy to manage that. It also seems to have  better performance by using a ring buffer for the inner subscriber to each source observable, as well as the list of subscriptions.

Patch-set: 13
Reviewer: Alex Bikfalvi <1002751@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0