Some refactoring of Importer classes for clarity
After trying to add a new translator and enhance the Neutron translation
I found the code a bit complex to follow. I'm proposing some
modifications here reducing some of the indirections, model hierarchies,
and naming trying to make the thing a bit more obvious for newcomers.

It also moves some of the classes from cluster to brain, following the
standard criteria that cluster hosts all stuff that is relevant to
interfaces exposed by the cluster to the rest of MidoNet services; brain
hosts the actual service implementations. This is increasingly important
to control the dependency graph accross modules.

Change-Id: If782ebfaa9a37ae2cb0815684d9f16adea4299c1
Signed-off-by: Galo Navarro <galo@midokura.com>