Modify getIp and getMtu to be shared among some tests
This patch modifies getIp and getMtu in DhcpTestCase to be shared with
other classes in org.midonet.midolman package, i.e.,
DhcpInterfaceMtuTestCase because the parameters specified in the command
are hardcoded and should be shared among other classes.

This also fixes the issue where the retrieval of MTU fails on Mac OSX when
it recovered from the sleep mode and the order of ifconfig output changed.

Change-Id: I3325f0e0bccea53a94f51923ee089d7e5f5c4374
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>