Implements await with FutureOps.
Waiting on a future such as one returned by Storage is a very common
operation. This implements "await()" method with FutureOps that waits
on a FutureOp to complete for a specified amount of time, 1 second by
default. Used with org.midonet.util.concurrent.toFutureOps, scala
Future is implicitly converted to FutureOp, and a client can wait for
its completion by simply calling await().

Change-Id: I6518ae82b0a0137d83dca1a1130be2ba61ed5663