Update patch set 2
Patch Set 2:

I thought that Galo's intention was that OneToOneNeutronTranslator was not very useful in practice as most of the translation wouldn't be so simple as one-to-one and that he wanted to avoid premature over-generalization. I don't mind having this, though. We can either wait till it becomes actually useful or put this now and remove later when it turned out indeed not very useful. What do you think?

Patch-set: 2
Label: Verified=0