Fixing the timeout in FlowStateStorageTest
A test in FlowStateStorageTest is executing a block of code surrounded
by an "eventually". Although one line in this block of code is using a
timeout of 3 seconds, the default timeout parameter for the "eventually"
block is 150 milliseconds. This sometimes results in a failed test.

This patch provides a custom timeout for the "eventually" block of 3
seconds.

Change-Id: I95b557944f2899caa9c0b19c7dde4308e5ed139c
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>